backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.6.0
aiosqlite==0.20.0
email-validator==2.2.0
supabase==2.9.1
python-dotenv==1.0.1
google-generativeai==0.3.2
python-multipart==0.0.20
psycopg[binary]==3.2.3
Pillow==11.0.0